requirement.txt
streamlit
google-generativeai
firebase-admin
twilio
python-dotenv
